Convert datasource config for local, qa, and prod (#1)
* Add application.yml to support switching datasources, define application
root context, and to set default spring profile (local). Updated
individual services to not lead with /api.

* Setup in-memory HSQL DB for Local Profile (default); Update DAO calls to
use new base jdbc dao class with NamedParameterJdbcTemplate for simpler
queries (more simplifying to come); use Long's intead of int/Integer for
consistency and future-proofing. Repackaged mysql folder on
src/main/resources

+++ src/main/java/svc/managers/CourtManager.java
@@ -14,7 +14,7 @@ public class CourtManager {
 	@Inject
 	private CourtDAO courtDAO;
 	
-	public Court GetCourtById(int courtId) {
+	public Court GetCourtById(Long courtId) {
 		return courtDAO.getByCourtId(courtId);
 	}
